+++ R/volcanoplot.R
@@ -12,9 +12,7 @@
 #' @export
 #'
 #' @examples
-#' mae <- hermes::multi_assay_experiment
-#' mae_data <- dataset("MAE", mae)
-#' data <- teal_data(mae_data)
+#' data <- teal_data(MAE = hermes::multi_assay_experiment)
 #' app <- init(
 #'   data = data,
 #'   modules = modules(
@@ -267,9 +265,7 @@ srv_g_volcanoplot <- function(id,
 #'   sample_tm_g_volcanoplot()
 #' }
 sample_tm_g_volcanoplot <- function() {
-  mae <- hermes::multi_assay_experiment
-  mae_data <- teal.data::dataset("MAE", mae)
-  data <- teal.data::teal_data(mae_data)
+  data <- teal.data::teal_data(MAE = hermes::multi_assay_experiment)
   app <- teal::init(
     data = data,
     modules = teal::modules(